Factorial limit fix
Fixed the issue with large numbers overflowing the stack, now it returns inf.

+++ src/math_lib.h
@@ -56,7 +56,7 @@ extern double div(const double num1, const double num2);
  * @param num number, from which is factorial made
  * @return factorial numbers from parameter
  */
-extern double fact(const double num);
+extern long double fact(const double num);
 
 /**
  * @brief Exponetiation with real base and natural exponent
